refactor: Refactor FieldArrays
Reordered field array files for testing
