+++ src/components/PublicationRequestFormSections/FieldArrays/ExternalRequestIdFieldArray/index.js
@@ -1,0 +1,1 @@
+export { default } from './ExternalRequestIdFieldArray';
