+++ src/components/PublicationRequestFormSections/FieldArrays/FundingFieldArray/index.js
@@ -1,0 +1,1 @@
+export { default } from './FundingFieldArray';
