+++ src/components/PublicationRequestFormSections/FieldArrays/IdentifiersFieldArray/index.js
@@ -1,0 +1,1 @@
+export { default } from './IdentifiersFieldArray';
