+++ src/components/PublicationRequestFormSections/FieldArrays/OtherEmailFieldArray/index.js
@@ -1,0 +1,1 @@
+export { default } from './OtherEmailFieldArray';
