+++ src/components/PublicationRequestFormSections/FieldArrays/PublicationStatusFieldArray/index.js
@@ -1,0 +1,1 @@
+export { default } from './PublicationStatusFieldArray';
