+++ src/components/PublicationRequestFormSections/FieldArrays/StreetAddressesFieldArray/index.js
@@ -1,0 +1,1 @@
+export { default } from './StreetAddressesFieldArray';
